Add files via upload
Game description : 
This Open Processing project is dedicated to making a single-player game that involves mouse indentation that permits players to interact with the game. The players control the speed of a ball on the screen with the click of a mouse. The ball starts in the center of the screen and initially moves in the right direction. When the ball hits the wall, it bounces back in the reverse direction. The aim of the game is to click on the moving ball while it moves from side to side to gain points. Simultaneously with every right-click, the speed of the ball increases progressively.  If he/she doesn't hit the ball, they lose 1 life point. The game expires when the user runs out of life points and the menu displays "Game Over".